Rename LibOvrIntegration to OvrIntegration
This includes namespaces, cmake variables, documentation, folders and
LibOvrIntegration.h
