+++ doc/namespaces.dox
@@ -43,18 +43,18 @@ path and link to `${MAGNUM_BULLETINTEGRATION_LIBRARIES}`. See
 @ref building-integration and @ref cmake-integration for more information.
 */
 
-/** @dir Magnum/LibOvrIntegration
- * @brief Namespace @ref Magnum::LibOvrIntegration
+/** @dir Magnum/OvrIntegration
+ * @brief Namespace @ref Magnum::OvrIntegration
  */
-/** @namespace Magnum::LibOvrIntegration
+/** @namespace Magnum::OvrIntegration
 @brief Integration with the Oculus SDK (LibOVR)
 
 Conversion of math classes, wrappers for most OVR CAPI structures and functions.
 
 This library depends on **Oculus SDK 0.6.0.1-beta** and is built if `WITH_OVR`
 is enabled when building Magnum Integration. To use this library, you need to
-request `LibOvr` component of `MagnumIntegration` package in CMake, add
-`${MAGNUM_LIBOVRINTEGRATION_INCLUDE_DIRS}` to include path and link to
-`${MAGNUM_LIBOVRINTEGRATION_LIBRARIES}`. See @ref building-integration and
+request `Ovr` component of `MagnumIntegration` package in CMake, add
+`${MAGNUM_OvrIntegration_INCLUDE_DIRS}` to include path and link to
+`${MAGNUM_OvrIntegration_LIBRARIES}`. See @ref building-integration and
 @ref cmake-integration for more information.
 */
